Moved fishing logic to a fishing controller
When player starts fishing, they will wait for a fish.  A progress bar
will show how long until the fishing operation is complete.  Once the
wait is finished, a fish will be caught, and displayed in the list of
fish caught.

If multiple of the same fish are caught, the badge will increment
instead of listing duplicate entries.

Letting go of the left mouse button will cancel the fishing event.

Once a fish is caught, another wait cycle will begin as long as the
player is still holding the left mouse button.